# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/media/{MEDIA_ID}": {
            "delete": {
                "tags": [
                    "Media"
                ],
                "summary": "Delete medium",
                "description": "Delete medium",
                "operationId": "destroy-medium",
                "parameters": [
                    {
                        "name": "MEDIA_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/memorials/{MEMORIAL_ID}/media": {
            "get": {
                "tags": [
                    "Memorials media"
                ],
                "summary": "Get all memorial media",
                "description": "Get all memorial media",
                "operationId": "get-memorial-media",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            },
            "post": {
                "tags": [
                    "Memorials media"
                ],
                "summary": "Upload media",
                "description": "Upload media to given memorial",
                "operationId": "upload-media",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "files",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "array[file]"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully uploaded"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
        "/api/v1/memorials/{MEMORIAL_ID}/video": {
            "get": {
                "tags": [
                    "Memorials media"
                ],
                "summary": "Get memorial videos",
                "description": "Get memorial videos",
                "operationId": "get-memorial-videos",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
        "/api/v1/memorials/{MEMORIAL_ID}/images": {
            "get": {
                "tags": [
                    "Memorials media"
                ],
                "summary": "Get memorial images",
                "description": "Get memorial images",
                "operationId": "get-memorial-images",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
        "/api/v1/memorials/media/{MEDIUM_ID}": {
            "delete": {
                "tags": [
                    "Memorials media"
                ],
                "summary": "Delete medium",
                "description": "Delete medium",
                "operationId": "destroy-memorial-medium",
                "parameters": [
                    {
                        "name": "MEDIUM_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/memorial/{MEMORIAL_ID}/media": {
            "delete": {
                "tags": [
                    "Memorials media"
                ],
                "summary": "Delete memorial media",
                "description": "Delete all memorial media",
                "operationId": "destroy-memorial-media",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/bouquets/{BOUQUET_ID}/media": {
            "post": {
                "tags": [
                    "Bouquets"
                ],
                "summary": "Upload medium",
                "description": "Upload medium to given bouquet",
                "operationId": "upload-medium",
                "parameters": [
                    {
                        "name": "BOUQUET_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "file",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MediaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
...
        "/api/v1/memorials/{MEMORIAL_ID}": {
            "get": {
                "tags": [
                    "Memorials"
                ],
                "summary": "Get memorial info",
                "description": "Get memorial info",
                "operationId": "get-memorial",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortMemorialResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Memorials"
                ],
                "summary": "Update memorial",
                "description": "Update memorial",
                "operationId": "update-memorial",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShortMemorialResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Memorials"
                ],
                "summary": "Change memorial status",
                "description": "Change memorial status",
                "operationId": "change-memorial-status",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/v1/memorial/{MEMORIAL_ID}": {
            "delete": {
                "tags": [
                    "Memorials"
                ],
                "summary": "Delete memorial",
                "description": "Delete all memorial",
                "operationId": "destroy-memorial",
                "parameters": [
                    {
                        "name": "MEMORIAL_ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MediaResource": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "dc9b0882-2f6f-4164-823d-ad2b0b8decb9"
                    },
                    "file_name": {
                        "type": "string",
                        "example": "FLS4yzPIjW1I1PkC1uymrulxBUcrXyHl7xiDuEpm.jpg"
                    },
                    "mime_type": {
                        "type": "string",
                        "example": "application/json"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MediaResource"
                }
            },
...
            "ShortMemorialResource": {
                "properties": {
                    "id": {
                        "type": "int8",
                        "example": "53"
                    },
                    "title": {
                        "type": "string",
                        "example": "Lorem ipsum"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "middle_name": {
                        "type": "string",
                        "example": "Fitzgerald"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "place_of_birth": {
                        "type": "string",
                        "example": "New York City"
                    },
                    "place_of_death": {
                        "type": "string",
                        "example": "Concord"
                    },
                    "day_of_birth": {
                        "type": "string",
                        "example": "2012-12-12T00:00:00.000000Z"
                    },
                    "day_of_death": {
                        "type": "string",
                        "example": "2012-12-13T00:00:00.000000Z"
                    },
                    "text": {
                        "type": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla vel libero leo."
                    },
                    "default": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "status": {
                        "type": "string",
                        "example": "published"
                    },
                    "user": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ShortUserResource"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ShortMemorialResource"
                }
            },
...
            "ShortUserResource": {
                "properties": {
                    "id": {
                        "type": "int8",
                        "example": "53"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "http://b-ouquet.com/storage/avatars/3gpbiSkdpCbyIADiyb0wuopeSVnbT0gBCJfZi0GL_avatar.jpg"
                    },
                    "memorial": {
                        "type": "int8",
                        "example": "35"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ShortUserResource"
                }
            }
        },
...
    }
}